Improvements based on PR review
