+++ common/ferrostar/src/navigation_controller/models.rs
@@ -4,9 +4,7 @@ use super::step_advance::conditions::ManualStepCondition;
 use super::step_advance::{SerializableStepAdvanceCondition, StepAdvanceCondition};
 use crate::algorithms::distance_between_locations;
 use crate::deviation_detection::{RouteDeviation, RouteDeviationTracking};
-use crate::models::{
-    Route, RouteStep, SpokenInstruction, UserLocation, VisualInstruction, Waypoint,
-};
+use crate::models::{RouteStep, SpokenInstruction, UserLocation, VisualInstruction, Waypoint};
 
 #[cfg(feature = "alloc")]
 use alloc::vec::Vec;
@@ -83,9 +81,9 @@ impl NavState {
 #[cfg_attr(feature = "wasm-bindgen", serde(rename_all = "camelCase"))]
 #[cfg_attr(feature = "wasm-bindgen", tsify(into_wasm_abi, from_wasm_abi))]
 pub struct SerializableNavState {
-    trip_state: TripState,
+    pub(crate) trip_state: TripState,
     // This has to be here because we actually do need to update the internal state that changes throughout navigation.
-    step_advance_condition: SerializableStepAdvanceCondition,
+    pub(crate) step_advance_condition: SerializableStepAdvanceCondition,
 }
 
 impl From<SerializableNavState> for NavState {
@@ -358,56 +356,4 @@ impl From<NavigationControllerConfig> for SerializableNavigationControllerConfig
             snapped_location_course_filtering: config.snapped_location_course_filtering,
         }
     }
-}
-
-/// An event that occurs during navigation.
-///
-/// This is used for the optional session recording / telemetry.
-#[derive(Clone, Serialize, Deserialize, Debug)]
-#[cfg_attr(feature = "uniffi", derive(uniffi::Record))]
-#[cfg_attr(feature = "wasm-bindgen", derive(Tsify))]
-pub struct NavigationRecordingEvent {
-    /// The timestamp of the event in milliseconds since Jan 1, 1970 UTC.
-    pub timestamp: i64,
-    /// Data associated with the event.
-    pub event_data: NavigationRecordingEventData,
-}
-
-impl NavigationRecordingEvent {
-    pub fn new(event_data: NavigationRecordingEventData) -> Self {
-        Self {
-            timestamp: Utc::now().timestamp_millis(),
-            event_data,
-        }
-    }
-
-    /// Create a [`NavigationRecordingEventData::StateUpdate`] event from a [`SerializableNavState`]
-    pub fn state_update(serializable_nav_state: SerializableNavState) -> Self {
-        Self::new(NavigationRecordingEventData::StateUpdate {
-            trip_state: serializable_nav_state.trip_state,
-            step_advance_condition: serializable_nav_state.step_advance_condition,
-        })
-    }
-
-    pub fn timestamp(&self) -> i64 {
-        self.timestamp
-    }
-}
-
-/// The event type.
-///
-/// For full replayability, we record things like rerouting, and not just location updates.
-#[derive(Clone, Serialize, Deserialize, Debug)]
-#[cfg_attr(feature = "uniffi", derive(uniffi::Enum))]
-#[cfg_attr(feature = "wasm-bindgen", derive(Tsify))]
-pub enum NavigationRecordingEventData {
-    StateUpdate {
-        trip_state: TripState,
-        step_advance_condition: SerializableStepAdvanceCondition,
-    },
-    // TODO: Figure out how to record re-routes.
-    RouteUpdate {
-        /// Updated route.
-        route: Route,
-    },
 }
